modify menu to link each page, add wave into banner, re-structure about us
