+++ components/Base/index.css
@@ -1,0 +1,11 @@
+.banner-content {
+    position: relative;
+}
+
+.banner-wave {
+    position: absolute;
+    overflow: hidden;
+    bottom: 0;
+    width: 100%;
+    height: 70%;
+}
