+++ pages/blog.js
@@ -1,3 +1,9 @@
+import MenuDefault from '../components/menu/menu.js'
+import FooterDefault from '../components/footer/footer.js'
 import BlogComponent from '../components/blogs/blog';
 
-export default BlogComponent;
+export default () => <div>
+    <MenuDefault activeItem='Blog'/>
+    <BlogComponent />
+    <FooterDefault />
+</div>
