+++ pages/blogs/index.js
@@ -1,3 +1,9 @@
+import MenuDefault from '../../components/menu/menu.js'
+import FooterDefault from '../../components/footer/footer.js'
 import BlogsComponent from '../../components/blogs';
 
-export default BlogsComponent;
+export default () => <div>
+    <MenuDefault activeItem='Blog'/>
+    <BlogsComponent />
+    <FooterDefault/>
+</div>
